go-travel/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.10.0
pydantic==2.5.3
anthropic==0.18.0
aiohttp==3.9.3
ortools>=9.0.0
